requirements.txt
selenium>=4.15.2
playwright>=1.40.0
pytest>=7.4.3
pytest-playwright>=0.4.3
webdriver-manager>=4.0.1
allure-pytest>=2.13.2
allure-python-commons>=2.13.2
Appium-Python-Client>=3.1.1
pytest-xdist>=3.5.0
pytest-rerunfailures>=13.0
python-dotenv>=1.0.0
assertpy>=1.1